work in progress creating sample data from superset examples
