+++ dataworkspace/dataworkspace/apps/example_data/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ExampleDataConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'dataworkspace.apps.example_data'
