+++ dataworkspace/dataworkspace/settings/base.py
@@ -103,6 +103,7 @@ INSTALLED_APPS = [
     'dataworkspace.apps.finder',
     'dynamic_models',
     'dataworkspace.apps.case_studies',
+
 ]
 
 MIDDLEWARE = [
